+++ templates/administrador/cadastro_adm.html
@@ -36,7 +36,7 @@
       <button type="submit">Criar Conta</button>
     </form>
     <div class="footer">
-      Já tem uma conta? <a href="/cadastro">Entrar</a>
+      Já tem uma conta? <a href="/login">Entrar</a>
     </div>
   </div>
   <!--Início das opções do canto-->
