+++ templates/administrador/perfil_adm.html
@@ -7,7 +7,7 @@
 </head>
 <body>
   <header class="topo">
-    <h1>Perfil da Admistrador</h1>
+    <h1>Perfil do Admistrador</h1>
     <!--Início das opções do canto-->
   </header>
   <div class="opcoes" onclick="toggleSidebar()">☰</div>
