+++ templates/empresa/cadastro_empresa.html
@@ -40,7 +40,7 @@
       <button type="submit">Cadastrar empresa</button>
     </form>
     <div class="footer">
-      Já tem uma conta? <a href="/cadastro">Entrar</a>
+      Já tem uma conta? <a href="/login">Entrar</a>
     </div>
   </div>
   <!--Início das opções do canto-->
