+++ templates/login.html
@@ -1,0 +1,43 @@
+<p></p>
+<!DOCTYPE html>
+<html lang="pt-BR">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>Login Site</title>
+    <link rel="stylesheet" href="/static/estilos/styles_login.css">
+</head>
+<body>
+    <div class= "Popup_Criar_conta">
+        <div class= "form-Popup Criar_conta">
+            <!--Logo ifce1 pq é a logo no Popup de "novo login", o Logo ifce2 é para o "Faça o login"-->
+            <i class="Logo-ifce1">
+            <img src="/static/imagens/Logo ifce.png" alt="Logo ifce1" class="Logo-ifce1">
+        </i>
+            <h2>Bem Vindo</h2>
+            <h3>Novo Login</h3>
+            <a href="/cadastro_usuario">
+                <button class="btn_criar_conta">Criar conta</button>
+            </a>
+        </div>
+        <div class="form-Popup Faça_login">
+            <!--###Logo-ifce2###-->
+            <i class="Logo-ifce2">
+                <img src="/static/imagens/Logo ifce.png" alt="Logo-ifce2" class="Logo-ifce2">
+            </i>
+            <h2>Faça Login</h2>
+           <form action="/login" method="post">
+            <!--Aqui o input para o email-->
+            <input type="text" name="nome" placeholder="Email" required>
+            <!--Aqui o input para a senha-->
+            <input type="password" name="senha" placeholder="Senha" required>
+            <div class="opcões">
+                <label> <input type="checkbox"></label>
+                <a href="#">Esqueceu senha?</a>
+            </div>
+            <button  type="submit" class="btn_entrar">Entrar</button>  
+           </form>
+    </div>
+</div>
+</body>
+</html>
