+++ templates/usuario/cadastro_usuario.html
@@ -36,7 +36,7 @@
         <button type="submit">Criar Conta</button>
       </form>
     <div class="footer">
-      Já tem uma conta? <a href="/cadastro">Entrar</a>
+      Já tem uma conta? <a href="/login">Entrar</a>
     </div>
   </div>
   <div class="popup-empresa">
